+++ bluetooth/macos/btsocket/btsocket.py
@@ -1,49 +1,50 @@
+import logging
 import time
-import socket as _socket
+import socket
 import threading
 import os
 import errno
-import types
 
 import objc
 import Foundation
 
-from bluetooth.btcommon import RFCOMM
+from bluetooth.btcommon import BluetoothError, Protocols
 from bluetooth.macos.loop import Loop
-import lightblue
-from lightblue import _IOBluetooth
-from lightblue import _lightbluecommon
-from lightblue import _macutil
+from bluetooth.macos import IOBluetooth
+from bluetooth.macos import util
+import bluetooth.address
 from lightblue._LightAquaBlue import BBServiceAdvertiser, BBBluetoothChannelDelegate
 
-#import sets     # python 2.3
+logger = logging.getLogger(__name__)
 
 try:
     SHUT_RD, SHUT_WR, SHUT_RDWR = \
-        _socket.SHUT_RD, _socket.SHUT_WR, _socket.SHUT_RDWR
+        socket.SHUT_RD, socket.SHUT_WR, socket.SHUT_RDWR
 except AttributeError:
     # python 2.3
     SHUT_RD, SHUT_WR, SHUT_RDWR = (0, 1, 2)
 
 
-def _getavailableport(proto):
+def _getavailableport(proto: Protocols):
     # Just advertise a service and see what channel it was assigned, then
     # stop advertising the service and return the channel.
     # It's a hacky way of doing it, but IOBluetooth doesn't seem to provide
     # functionality for just getting an available channel.
 
-    if proto == _lightbluecommon.RFCOMM:
+    if proto == Protocols.RFCOMM:
         try:
             result, channelID, servicerecord = BBServiceAdvertiser.addRFCOMMServiceDictionary_withName_UUID_channelID_serviceRecord_(BBServiceAdvertiser.serialPortProfileDictionary(), "DummyService", None, None, None)
         except:
             result, channelID, servicerecord = BBServiceAdvertiser.addRFCOMMServiceDictionary_withName_UUID_channelID_serviceRecord_(BBServiceAdvertiser.serialPortProfileDictionary(), "DummyService", None)
-        if result != _macutil.kIOReturnSuccess:
-            raise _lightbluecommon.BluetoothError(result, \
-                "Could not retrieve an available service channel")
+        
+        if result != util.kIOReturnSuccess:
+            raise BluetoothError(f"Could not retrieve an available service channel ({result})")
+        
         result = BBServiceAdvertiser.removeService_(servicerecord)
-        if result != _macutil.kIOReturnSuccess:
-            raise _lightbluecommon.BluetoothError(result, \
-                "Could not retrieve an available service channel")
+        
+        if result != util.kIOReturnSuccess:
+            raise BluetoothError(f"Could not retrieve an available service channel ({result})")
+        
         return channelID
 
     else:
@@ -55,32 +56,27 @@ def _checkaddrpair(address, checkbtaddr=True):
     # (for binding to a server address)
 
     if not isinstance(address, tuple):
-        raise TypeError("address must be (address, port) tuple, was %s" % \
-            type(address))
+        raise TypeError(f"address must be (address, port) tuple, was {type(address)}")
 
     if len(address) != 2:
-        raise TypeError("address tuple must have 2 items (has %d)" % \
-            len(address))
+        raise TypeError(f"address tuple must have 2 items (has {len(address)})")
 
     if not isinstance(address[0], str):
-        raise TypeError("address host value must be string, was %s" % \
-            type(address[0]))
+        raise TypeError(f"address host value must be string, was {type(address[0])}")
 
     if checkbtaddr:
-        if not _lightbluecommon._isbtaddr(address[0]):
-            raise TypeError("address '%s' is not a bluetooth address" % \
-                address[0])
+        if not bluetooth.address.validate(address[0]):
+            raise TypeError(f"invalid address: {address[0]}")
 
     if not isinstance(address[1], int):
-        raise TypeError("address port value must be int, was %s" % \
-            type(address[1]))
+        raise TypeError(f"address port value must be int, was {type(address[1])}")
 
 
 # from std lib socket module
 class _closedsocket:
     __slots__ = []
     def _dummy(*args):
-        raise _socket.error(errno.EBADF, 'Bad file descriptor')
+        raise socket.error(errno.EBADF, 'Bad file descriptor')
     send = recv = sendto = recvfrom = __getattr__ = _dummy
 
 
@@ -112,8 +108,8 @@ class _ByteQueue:
         return result
 
 
-#class _SocketWrapper(_socket._socketobject):
-class _SocketWrapper:
+#class socketWrapper(socket.socketobject):
+class socketWrapper:
     """
     A Bluetooth socket object has the same interface as a socket object from
     the Python standard library <socket> module. It also uses the same
@@ -151,12 +147,10 @@ class _SocketWrapper:
 
     def accept(self):
         sock, addr = self._sock.accept()
-        return _SocketWrapper(sock), addr
-    accept.__doc__ = _lightbluecommon._socketdocs["accept"]
+        return socketWrapper(sock), addr
 
     def dup(self):
-        return _SocketWrapper(self._sock)
-    dup.__doc__ = _lightbluecommon._socketdocs["dup"]
+        return socketWrapper(self._sock)
 
     def close(self):
         self._sock.close()
@@ -169,12 +163,10 @@ class _SocketWrapper:
             lightblue.stopadvertise(self)
         except:
             pass
-    close.__doc__ = _lightbluecommon._socketdocs["close"]
 
     def makefile(self, mode='r', bufsize=-1):
         # use std lib socket's _fileobject
-        return _socket._fileobject(self._sock, mode, bufsize)
-    makefile.__doc__ = _lightbluecommon._socketdocs["makefile"]
+        return socket._fileobject(self._sock, mode, bufsize)
 
     # delegate all other method calls to internal sock obj
     def __getattr__(self, attr):
@@ -183,9 +175,8 @@ class _SocketWrapper:
 
 # internal _sock object for RFCOMM and L2CAP sockets
 class _BluetoothSocket:
-
-    _boundports = { _lightbluecommon.L2CAP: set(),
-                    _lightbluecommon.RFCOMM: set() }
+    _boundports = { Protocols.L2CAP: set(),
+                    Protocols.RFCOMM: set() }
 
     # conn is the associated _RFCOMMConnection or _L2CAPConnection
     def __init__(self, conn):
@@ -214,9 +205,9 @@ class _BluetoothSocket:
 
     def accept(self):
         if not self.__isbound():
-            raise _socket.error('Socket not bound')
+            raise socket.error('Socket not bound')
         if not self.__islistening():
-            raise _socket.error('Socket must be listening first')
+            raise socket.error('Socket must be listening first')
 
         def clientconnected():
             return len(self.__queuedchannels) > 0
@@ -233,23 +224,23 @@ class _BluetoothSocket:
             self.__queuedchannels_lock.release()
 
         # return (new-socket, addr) pair using the new channel
-        newconn = _SOCKET_CLASSES[self.__conn.proto](newchannel)
-        sock = _SocketWrapper(_BluetoothSocket(newconn))
+        newconn = socket_CLASSES[self.__conn.proto](newchannel)
+        sock = socketWrapper(_BluetoothSocket(newconn))
         sock.__startevents()
         return (sock, sock.getpeername())
 
     def bind(self, address):
         _checkaddrpair(address, False)
         if self.__isbound():
-            raise _socket.error('Socket is already bound')
+            raise socket.error('Socket is already bound')
         elif self.__isconnected():
-            raise _socket.error("Socket is already connected, cannot be bound")
+            raise socket.error("Socket is already connected, cannot be bound")
 
-        if self.__conn.proto == _lightbluecommon.L2CAP:
+        if self.__conn.proto == Protocols.L2CAP:
             raise NotImplementedError("L2CAP server sockets not currently supported")
 
         if address[1] != 0:
-            raise _socket.error("must bind to port 0, other ports not supported on Mac OS X")
+            raise socket.error("must bind to port 0, other ports not supported on Mac OS X")
         address = (address[0], _getavailableport(self.__conn.proto))
 
         # address must be either empty string or local device address
@@ -260,12 +251,12 @@ class _BluetoothSocket:
             except:
                 localaddr = None
             if localaddr is None or address[0] != localaddr:
-                raise _socket.error(
+                raise socket.error(
                     errno.EADDRNOTAVAIL, os.strerror(errno.EADDRNOTAVAIL))
 
         # is this port already in use?
         if address[1] in self._boundports[self.__conn.proto]:
-            raise _socket.error(errno.EADDRINUSE, os.strerror(errno.EADDRINUSE))
+            raise socket.error(errno.EADDRINUSE, os.strerror(errno.EADDRINUSE))
 
         self._boundports[self.__conn.proto].add(address[1])
         self.__port = address[1]
@@ -297,13 +288,13 @@ class _BluetoothSocket:
 
     def connect(self, address):
         if self.__isbound():
-            raise _socket.error("Can't connect, socket has been bound")
+            raise socket.error("Can't connect, socket has been bound")
         elif self.__isconnected():
-            raise _socket.error("Socket is already connected")
+            raise socket.error("Socket is already connected")
         _checkaddrpair(address)
 
         # open a connection to device
-        self.__remotedevice = _IOBluetooth.IOBluetoothDevice.withAddressString_(address[0])
+        self.__remotedevice = IOBluetooth.IOBluetoothDevice.withAddressString_(address[0])
 
         if not self.__remotedevice.isConnected():
             if self.__timeout is None:
@@ -312,15 +303,15 @@ class _BluetoothSocket:
                 result = self.__remotedevice.openConnection_withPageTimeout_authenticationRequired_(
                     None, self.__timeout*1000, False)
 
-            if result != _macutil.kIOReturnSuccess:
-                if result == _macutil.kBluetoothHCIErrorPageTimeout:
+            if result != util.kIOReturnSuccess:
+                if result == util.kBluetoothHCIErrorPageTimeout:
                     if self.__timeout == 0:
-                        raise _socket.error(errno.EAGAIN,
+                        raise socket.error(errno.EAGAIN,
                             "Resource temporarily unavailable")
                     else:
-                        raise _socket.timeout("connect timed out")
+                        raise socket.timeout("connect timed out")
                 else:
-                    raise _socket.error(result,
+                    raise socket.error(result,
                         "Cannot connect to %s, can't open connection." \
                                                             % str(address[0]))
 
@@ -329,11 +320,11 @@ class _BluetoothSocket:
         result = self.__conn.connect(self.__remotedevice, address[1],
                 self.__eventlistener)   # pass listener as cocoa delegate
 
-        if result != _macutil.kIOReturnSuccess:
+        if result != util.kIOReturnSuccess:
             self.__remotedevice.closeConnection()
             self.__stopevents()
             self.__eventlistener = None
-            raise _socket.error(result,
+            raise socket.error(result,
                     "Cannot connect to %d on %s" % (address[1], address[0]))
             return
 
@@ -344,18 +335,18 @@ class _BluetoothSocket:
     def connect_ex(self, address):
         try:
             self.connect(address)
-        except _socket.error as err:
+        except socket.error as err:
             if len(err.args) > 1:
                 return err.args[0]
             else:
                 # there's no error code, just a message, so this error wasn't
                 # from a system call -- so re-raise the exception
-                raise _socket.error(err)
+                raise socket.error(err)
         return 0
 
     def getpeername(self):
         self.__checkconnected()
-        addr = _macutil.formatdevaddr(self.__remotedevice.getAddressString())
+        addr = util.formatdevaddr(self.__remotedevice.getAddressString())
         return (addr, self._getport())
 
     def getsockname(self):
@@ -370,7 +361,7 @@ class _BluetoothSocket:
             return
 
         if not self.__isbound():
-            raise _socket.error('Socket not bound')
+            raise socket.error('Socket not bound')
         if not isinstance(backlog, int):
             raise TypeError("backlog must be int, was %s" % type(backlog))
         if backlog < 0:
@@ -384,7 +375,7 @@ class _BluetoothSocket:
 
     def _isclosed(self):
         # isOpen() check doesn't work for incoming (server-spawned) channels
-        if (self.__conn.proto == _lightbluecommon.RFCOMM and
+        if (self.__conn.proto == Protocols.RFCOMM and
                 self.__conn.channel is not None and
                 not self.__conn.channel.isIncoming()):
             return not self.__conn.channel.isOpen()
@@ -411,7 +402,7 @@ class _BluetoothSocket:
 
         if self._isclosed():
             if len(self.__incomingdata) == 0:
-                raise _socket.error(errno.ECONNRESET,
+                raise socket.error(errno.ECONNRESET,
                                     os.strerror(errno.ECONNRESET))
             return self.__incomingdata.read(bufsize)
 
@@ -424,7 +415,7 @@ class _BluetoothSocket:
 
         # other side closed connection while waiting?
         if self._isclosed() and len(self.__incomingdata) == 0:
-            raise _socket.error(errno.ECONNRESET, os.strerror(errno.ECONNRESET))
+            raise socket.error(errno.ECONNRESET, os.strerror(errno.ECONNRESET))
 
         return self.__incomingdata.read(bufsize)
 
@@ -447,7 +438,7 @@ class _BluetoothSocket:
         if not isinstance(data, (bytes, bytearray)):
             raise TypeError("data must be bytes, was %s" % type(data))
         if self.__commstate in (SHUT_WR, SHUT_RDWR):
-            raise _socket.error(errno.EPIPE, os.strerror(errno.EPIPE))
+            raise socket.error(errno.EPIPE, os.strerror(errno.EPIPE))
         self.__checkconnected()
 
         # do setup for if sock is in non-blocking mode
@@ -456,10 +447,10 @@ class _BluetoothSocket:
                 # in non-blocking mode
                 # isTransmissionPaused() is not available for L2CAP sockets,
                 # what to do for that?
-                if self.__conn.proto == _lightbluecommon.RFCOMM and \
+                if self.__conn.proto == Protocols.RFCOMM and \
                         self.__conn.channel.isTransmissionPaused():
                     # sending data now will block
-                    raise _socket.error(errno.EAGAIN,
+                    raise socket.error(errno.EAGAIN,
                         "Resource temporarily unavailable")
             elif self.__timeout > 0:
                 # non-blocking with timeout
@@ -472,7 +463,7 @@ class _BluetoothSocket:
         while bytesleft > 0:
             if self.__timeout is not None and self.__timeout > 0:
                 if time.time() - starttime > self.__timeout:
-                    raise _socket.timeout("send timed out")
+                    raise socket.timeout("send timed out")
 
             # write the data to the channel (only the allowed amount)
             # the method/selector is the same for L2CAP and RFCOMM channels
@@ -487,8 +478,8 @@ class _BluetoothSocket:
             # normal tcp sockets don't seem to actually error on the first
             # send() after a connection has broken; if you try a second time,
             # then you get the (32, 'Broken pipe') socket.error
-            if result != _macutil.kIOReturnSuccess:
-                raise _socket.error(result, "Error sending data")
+            if result != util.kIOReturnSuccess:
+                raise socket.error(result, "Error sending data")
 
             bytesleft -= sendbytecount
             writebuf = writebuf[sendbytecount:] # remove the data just sent
@@ -524,13 +515,13 @@ class _BluetoothSocket:
     def getsockopt(self, level, optname, buflen=0):
         # see what options on Linux+s60
         # possibly have socket security option.
-        raise _socket.error(
+        raise socket.error(
             errno.ENOPROTOOPT, os.strerror(errno.ENOPROTOOPT))
 
     def setsockopt(self, level, optname, value):
         # see what options on Linux+s60
         # possibly have socket security option.
-        raise _socket.error(
+        raise socket.error(
             errno.ENOPROTOOPT, os.strerror(errno.ENOPROTOOPT))
 
     def setblocking(self, flag):
@@ -554,7 +545,7 @@ class _BluetoothSocket:
 
     def shutdown(self, how):
         if how not in (SHUT_RD, SHUT_WR, SHUT_RDWR):
-            raise _socket.error(22, "Invalid argument")
+            raise socket.error(22, "Invalid argument")
         self.__commstate = how
 
     # This method is called from outside this file.
@@ -563,7 +554,7 @@ class _BluetoothSocket:
             return self.__conn.getport()
         if self.__isbound():
             return self.__port
-        raise _lightbluecommon.BluetoothError("socket is neither connected nor bound")
+        raise BluetoothError("socket is neither connected nor bound")
 
     # This method is called from outside this file.
     def _getchannel(self):
@@ -611,18 +602,18 @@ class _BluetoothSocket:
                 Loop().wait_blocking(0.02)
                 if not stopwaiting():
                     # trying to perform operation now would block
-                    raise _socket.error(errno.EAGAIN, os.strerror(errno.EAGAIN))
+                    raise socket.error(errno.EAGAIN, os.strerror(errno.EAGAIN))
             else:
                 # block and wait until we get data, or time out
                 if not Loop().wait_until(stopwaiting, self.__timeout):
-                    raise _socket.timeout(timeoutmsg)
+                    raise socket.timeout(timeoutmsg)
 
     def __createlistener(self):
         if self.__isbound():
-            return _ChannelServerEventListener.alloc().initWithDelegate_port_protocol_(self,
+            return ChannelServerEventListener.alloc().initWithDelegate_port_protocol_(self,
                     self._getport(), self.__conn.proto)
         else:
-            listener = _ChannelEventListener.alloc().initWithDelegate_(self)
+            listener = ChannelEventListener.alloc().initWithDelegate_(self)
             if self.__conn.channel is not None:
                 self.__conn.channel.setDelegate_(listener.delegate())
                 listener.registerclosenotif(self.__conn.channel)
@@ -631,7 +622,7 @@ class _BluetoothSocket:
     # should not call this if connect() has been called to connect this socket
     def __startevents(self):
         if self.__eventlistener is not None:
-            raise _lightbluecommon.BluetoothError("socket already listening")
+            raise BluetoothError("socket already listening")
         self.__eventlistener = self.__createlistener()
 
     def __stopevents(self):
@@ -644,7 +635,7 @@ class _BluetoothSocket:
     def __checkconnected(self):
         if not self._sock.isconnected():  # i.e. is connected, non-server socket
             # not connected, raise "socket not connected"
-            raise _socket.error(errno.ENOTCONN, os.strerror(errno.ENOTCONN))
+            raise socket.error(errno.ENOTCONN, os.strerror(errno.ENOTCONN))
 
     # returns whether socket is a bound server socket
     def __isbound(self):
@@ -656,20 +647,12 @@ class _BluetoothSocket:
     def __checkconnected(self):
         if not self.__isconnected():
             # not connected, raise "socket not connected"
-            raise _socket.error(errno.ENOTCONN, os.strerror(errno.ENOTCONN))
-
-    # set method docstrings
-    definedmethods = locals()   # i.e. defined methods in _SocketWrapper
-    for name, doc in list(_lightbluecommon._socketdocs.items()):
-        try:
-            definedmethods[name].__doc__ = doc
-        except KeyError:
-            pass
+            raise socket.error(errno.ENOTCONN, os.strerror(errno.ENOTCONN))
 
 
 class _RFCOMMConnection:
 
-    proto = _lightbluecommon.RFCOMM
+    proto = Protocols.RFCOMM
 
     def __init__(self, channel=None):
         # self.channel is accessed by _BluetoothSocket parent
@@ -678,11 +661,11 @@ class _RFCOMMConnection:
     def connect(self, device, port, listener):
         # open RFCOMM channel (should timeout actually apply to opening of
         # channel as well? if so need to do timeout with async callbacks)
-        print("doing connect", device, port, listener)
+        logger.info(f"connecting RFCOMM: {device} {port} {listener}")
         result, self.channel = device.openRFCOMMChannelSync_withChannelID_delegate_(None, port, listener.delegate())
-        print("connect", result)
+        logger.info(f"connecting result: {result}")
         time.sleep(1)
-        if result == _macutil.kIOReturnSuccess:
+        if result == util.kIOReturnSuccess:
             self.channel.setDelegate_(listener.delegate())
             listener.registerclosenotif(self.channel)
         else:
@@ -691,7 +674,7 @@ class _RFCOMMConnection:
 
     def write(self, data):
         if self.channel is None:
-            raise _socket.error("socket not connected")
+            raise socket.error("socket not connected")
         return \
             BBBluetoothChannelDelegate.synchronouslyWriteData_toRFCOMMChannel_(
                 Foundation.NSData.alloc().initWithBytes_length_(data, len(data)),
@@ -706,7 +689,7 @@ class _RFCOMMConnection:
 
 class _L2CAPConnection:
 
-    proto = _lightbluecommon.L2CAP
+    proto = Protocols.L2CAP
 
     def __init__(self, channel=None):
         # self.channel is accessed by _BluetoothSocket parent
@@ -718,7 +701,7 @@ class _L2CAPConnection:
             result, self.channel = device.openL2CAPChannelSync_withPSM_delegate_(None, port, listener.delegate())
         except TypeError:
             result, self.channel = device.openL2CAPChannelSync_withPSM_delegate_(port, listener.delegate())
-        if result == _macutil.kIOReturnSuccess:
+        if result == util.kIOReturnSuccess:
             self.channel.setDelegate_(listener.delegate())
             listener.registerclosenotif(self.channel)
         else:
@@ -727,7 +710,7 @@ class _L2CAPConnection:
 
     def write(self, data):
         if self.channel is None:
-            raise _socket.error("socket not connected")
+            raise socket.error("socket not connected")
         return \
             BBBluetoothChannelDelegate.synchronouslyWriteData_toL2CAPChannel_(
                 bytes(data), self.channel)
@@ -739,7 +722,7 @@ class _L2CAPConnection:
         return self.channel.getPSM()
 
 
-class _ChannelEventListener(Foundation.NSObject):
+class ChannelEventListener(Foundation.NSObject):
     """
     Uses a BBBluetoothChannelDelegate to listen for events on an
     IOBluetoothRFCOMMChannel or IOBluetoothL2CAPChannel, and makes callbacks to
@@ -813,7 +796,7 @@ class _ChannelEventListener(Foundation.NSObject):
     channelClosed_ = objc.selector(channelClosed_, signature=b"v@:@")
 
 
-class _ChannelServerEventListener(Foundation.NSObject):
+class ChannelServerEventListener(Foundation.NSObject):
     """
     Listens for server-specific events on a RFCOMM or L2CAP channel (i.e. when a
     client connects) and makes callbacks to a specified object when events
@@ -837,13 +820,13 @@ class _ChannelServerEventListener(Foundation.NSObject):
         self.__cb_obj = cb_obj
         self.__usernotif = None
 
-        if proto == _lightbluecommon.RFCOMM:
-            usernotif = _IOBluetooth.IOBluetoothRFCOMMChannel.registerForChannelOpenNotifications_selector_withChannelID_direction_(self, "newChannelOpened:channel:", port, _macutil.kIOBluetoothUserNotificationChannelDirectionIncoming)
-        elif proto == _lightbluecommon.L2CAP:
-            usernotif = _IOBluetooth.IOBluetoothL2CAPChannel.registerForChannelOpenNotifications_selector_withPSM_direction_(self, "newChannelOpened:channel:", port, _macutil.kIOBluetoothUserNotificationChannelDirectionIncoming)
+        if proto == Protocols.RFCOMM:
+            usernotif = IOBluetooth.IOBluetoothRFCOMMChannel.registerForChannelOpenNotifications_selector_withChannelID_direction_(self, "newChannelOpened:channel:", port, util.kIOBluetoothUserNotificationChannelDirectionIncoming)
+        elif proto == Protocols.L2CAP:
+            usernotif = IOBluetooth.IOBluetoothL2CAPChannel.registerForChannelOpenNotifications_selector_withPSM_direction_(self, "newChannelOpened:channel:", port, util.kIOBluetoothUserNotificationChannelDirectionIncoming)
 
         if usernotif is None:
-            raise _socket.error("Unable to register for channel-" + \
+            raise socket.error("Unable to register for channel-" + \
                 "opened notifications on server socket on channel/PSM %d" % \
                 port)
         self.__usernotif = usernotif
@@ -874,27 +857,26 @@ class _ChannelServerEventListener(Foundation.NSObject):
 # -----------------------------------------------------------
 
 # protocol-specific classes
-_SOCKET_CLASSES = { _lightbluecommon.RFCOMM: _RFCOMMConnection,
-                    _lightbluecommon.L2CAP:  _L2CAPConnection }
+socket_CLASSES = { Protocols.RFCOMM: _RFCOMMConnection,
+                    Protocols.L2CAP:  _L2CAPConnection }
 
 def _getsocketobject(proto):
-    if proto not in list(_SOCKET_CLASSES.keys()):
+    if proto not in list(socket_CLASSES.keys()):
         raise ValueError("Unknown socket protocol, must be L2CAP or RFCOMM")
-    return _SocketWrapper(_BluetoothSocket(_SOCKET_CLASSES[proto]()))
+    return socketWrapper(_BluetoothSocket(socket_CLASSES[proto]()))
 
 
 class BluetoothSocket:
 
-    def __init__(self, proto=RFCOMM, _sock=None):
+    def __init__(self, proto: Protocols=Protocols.RFCOMM, _sock=None):
+        if proto != Protocols.RFCOMM:
+            raise NotImplementedError("Unsupported protocol")
+
         if _sock is None:
-            _proto=_lightbluecommon.RFCOMM
-            _sock = _getsocketobject(_proto)
+            _sock = _getsocketobject(proto)
         self._sock = _sock
 
-        if proto != RFCOMM:
-            # name the protocol
-            raise NotImplementedError("Not supported protocol")
-        self._proto = lightblue.RFCOMM
+        self._proto = proto
         self._addrport = None
 
     def _getport(self):
